Fix communications migration

+++ db/migrate/20260609085813_remove_kind_from_communications.rb
@@ -1,5 +1,5 @@
 class RemoveKindFromCommunications < ActiveRecord::Migration[8.1]
   def change
-    remove_column :communications, :kind, :string
+    remove_column :communications, :kind, :string if column_exists?(:communications, :kind)
   end
 end
